Move non-ui into a separate project
