+++ XCalendar.Core/Converters/EnumToStringConverter.cs
@@ -1,0 +1,25 @@
+﻿using System;
+
+namespace XCalendar.Core.Converters
+{
+    public static class EnumToStringConverter
+    {
+        public static object Convert(object value)
+        {
+            try
+            {
+                return Enum.GetName(value.GetType(), value);
+            }
+            
+            catch
+            {
+                return value?.ToString();
+            }
+        }
+
+        public static object ConvertBack(object value, object parameter)
+        {
+            return Enum.Parse(Type.GetType((string)parameter, true), (string)value);
+        }
+    }
+}
