+++ XCalendar.Core/Extensions/DayOfWeekExtensionscs.cs
@@ -1,10 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
-using System.Globalization;
-using System.Linq;
 
-namespace XCalendar.Extensions
+namespace XCalendar.Core.Extensions
 {
     public static class DayOfWeekExtensions
     {
