+++ XCalendar/Converters/EnumToStringConverter.cs
@@ -2,26 +2,18 @@
 using System.Globalization;
 using Xamarin.Forms;
 
-namespace XCalendar.Converters
+namespace XCalendar.Forms.Converters
 {
     public class EnumToStringConverter : IValueConverter
     {
         public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
         {
-            try
-            {
-                return Enum.GetName(value.GetType(), value);
-            }
-            
-            catch
-            {
-                return value?.ToString();
-            }
+            return Core.Converters.EnumToStringConverter.Convert(value);
         }
 
         public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
         {
-            return Enum.Parse(Type.GetType((string)parameter, true), (string)value);
+            return Core.Converters.EnumToStringConverter.ConvertBack(value, parameter);
         }
     }
 }
