+++ XCalendar/Converters/StringCharLimitConverter.cs
@@ -2,34 +2,13 @@
 using System.Globalization;
 using Xamarin.Forms;
 
-namespace XCalendar.Converters
+namespace XCalendar.Forms.Converters
 {
     public class StringCharLimitConverter : IValueConverter
     {
         public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
         {
-            try
-            {
-                string StringValue = (string)value;
-                int TargetLength = System.Convert.ToInt32(parameter);
-
-                if (TargetLength == 0)
-                {
-                    return "";
-                }
-                else if (TargetLength >= StringValue.Length)
-                {
-                    return StringValue;
-                }
-                else
-                {
-                    return StringValue.Substring(0, TargetLength);
-                }
-            }
-            catch
-            {
-                return "";
-            }
+            return Core.Converters.StringCharLimitConverter.Convert(value, parameter);
         }
 
         public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
