+++ XCalendar/Views/FramedLabel.xaml
@@ -1,9 +1,9 @@
 ﻿<?xml version="1.0" encoding="UTF-8" ?>
 <Frame
-    x:Class="XCalendar.FramedLabel"
+    x:Class="XCalendar.Forms.Views.FramedLabel"
     xmlns="http://xamarin.com/schemas/2014/forms"
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-    xmlns:xc="clr-namespace:XCalendar"
+    xmlns:xc="clr-namespace:XCalendar.Forms.Views"
     x:Name="MainFramedLabel"
     x:DataType="{x:Type xc:FramedLabel}">
