+++ XCalendar/Views/FramedLabel.xaml.cs
@@ -1,7 +1,7 @@
 ﻿using Xamarin.Forms;
 using Xamarin.Forms.Xaml;
 
-namespace XCalendar
+namespace XCalendar.Forms.Views
 {
     [XamlCompilation(XamlCompilationOptions.Compile)]
     public partial class FramedLabel : Frame
