+++ XCalendar/XCalendar.Forms.csproj
@@ -26,15 +26,19 @@
     <PackageReference Include="Xamarin.CommunityToolkit" Version="1.3.0" />
     <PackageReference Include="Xamarin.Forms" Version="5.0.0.2196" />
   </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\XCalendar.Core\XCalendar.Core.csproj" />
+  </ItemGroup>
 	
   <ItemGroup>
-    <EmbeddedResource Update="CalendarDayView.xaml">
+    <EmbeddedResource Update="Views\CalendarDayView.xaml">
       <Generator>MSBuild:Compile</Generator>
     </EmbeddedResource>
-    <EmbeddedResource Update="CalendarView.xaml">
+    <EmbeddedResource Update="Views\CalendarView.xaml">
       <Generator>MSBuild:Compile</Generator>
     </EmbeddedResource>
-    <EmbeddedResource Update="FramedLabel.xaml">
+    <EmbeddedResource Update="Views\FramedLabel.xaml">
       <Generator>MSBuild:Compile</Generator>
     </EmbeddedResource>
   </ItemGroup>
