+++ XCalendarSample/XCalendarSample/Popups/DatePickerDialogPopup.xaml
@@ -4,9 +4,9 @@
     xmlns="http://xamarin.com/schemas/2014/forms"
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
     xmlns:System="clr-namespace:System;assembly=mscorlib"
-    xmlns:xc="clr-namespace:XCalendar;assembly=XCalendar"
-    xmlns:xcConverters="clr-namespace:XCalendar.Converters;assembly=XCalendar"
-    xmlns:xcModels="clr-namespace:XCalendar.Models;assembly=XCalendar"
+    xmlns:xc="clr-namespace:XCalendar.Forms.Views;assembly=XCalendar.Forms"
+    xmlns:xcConverters="clr-namespace:XCalendar.Forms.Converters;assembly=XCalendar.Forms"
+    xmlns:xcModels="clr-namespace:XCalendar.Core.Models;assembly=XCalendar.Core"
     xmlns:xct="http://xamarin.com/schemas/2020/toolkit"
     x:Name="This"
     x:TypeArguments="System:DateTime"
