+++ XCalendarSample/XCalendarSample/Views/EventCalendarExamplePage.xaml
@@ -6,9 +6,9 @@
     xmlns:Converters="clr-namespace:XCalendarSample.Converters"
     xmlns:Models="clr-namespace:XCalendarSample.Models"
     xmlns:ViewModels="clr-namespace:XCalendarSample.ViewModels"
-    xmlns:xc="clr-namespace:XCalendar;assembly=XCalendar"
-    xmlns:xcConverters="clr-namespace:XCalendar.Converters;assembly=XCalendar"
-    xmlns:xcModels="clr-namespace:XCalendar.Models;assembly=XCalendar"
+    xmlns:xc="clr-namespace:XCalendar.Forms.Views;assembly=XCalendar.Forms"
+    xmlns:xcConverters="clr-namespace:XCalendar.Forms.Converters;assembly=XCalendar.Forms"
+    xmlns:xcModels="clr-namespace:XCalendar.Core.Models;assembly=XCalendar.Core"
     xmlns:xct="http://xamarin.com/schemas/2020/toolkit"
     x:Name="This"
     Title="Event Calendar"
